Functions and methods
Functions are an on demand machine that preforms a specific action.  It can be written to add coins or lumber to your resource pool.  It can be written to keep a closed until the player opens it by turning the knob.  They are repeatable but limited to the scope they was written for.